solved IO texture format issues in low-level processes
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/cortical_surface/cortical_surface-gpl/branches/bug_fix@72586 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ brainvisa/toolboxes/cortical_surface/processes/anatomy/tools/sulcalPitsLabelizationFromSulci.py
@@ -99,7 +99,7 @@ def execution( self, context ):
     context.write('labelling the pits')
     pits = np.array( pits_tex[0] )
     pits_ind = np.where(pits)[0]
-    pits[pits_ind] = basins_labeled[pits_ind]
+    pits[pits_ind] = np.uint16(basins_labeled[pits_ind])
     tex_out = aims.TimeTexture_S16()
     tex_out[0].assign(pits)
     aims.write(tex_out, self.labeled_pits_texture.fullPath())
